InfluxDB Outputter now manages the timestamp on events
Event objects have always had a `time` attribute but it wasn't being
used by the influxdb outputter, rather it relied on influxdb setting
the timestamp to the server time.

This could be problematic when using asynchronous writer in influxdb
client as the timestamp could skew from the time the event was
actually produced.

This change generates an event timestamp based on the time in the
Event object modified to match the time precision set in the influxdb client.